+++ lib/listiller.ex
@@ -1,33 +1,33 @@
 defmodule PhoenixLiveViewExt.Listiller do
   @moduledoc ~S'''
-  Helps LiveView optimally insert, update and delete list elements with replacing the entire list only when absolutely
+  Helps LiveView optimally insert, update and delete list elements while replacing the entire list only when absolutely
   necessary (i.e. when initialized).
 
-  Transforms new state relative to the old one into a list of LiveComponent assigns so as to leave out the assigns for
-  any listed component instances for which the new state does not result changed after the transformation. In other
+  Transforms new state relative to the old one into a list of LiveComponent assigns so as to leave out the assigns of
+  all listed component instances for which the new state does not result changed after the transformation. In other
   words, it diffs the list of LiveComponent assigns resulting from the state-to-assigns transformation (the components
   of which list are typically rendered as children of an appended or prepended container element).
 
-  The purpose of Listiller is to allow for optimal implementation of appended/prepended container lists where only the
-  list elements that have actually changed are sent from the server to the JS client over the wire and patched when
-  client-side LiveView traverses the DOM elements. While this can be manually optimized for individual component
-  instances by sending changes via `LiveView.send_update/3` such an approach will not suffice in cases with multiple
-  component changes and more importantly, in the cases where there are elements required for insertion to or deletion
-  from the container list.
+  The purpose of Listiller is to allow for an optimal implementation of appended/prepended container lists where only
+  the list elements that actually change get sent from the server to the JS client over the wire and patched when
+  client-side LiveView traverses the DOM elements. While this could be otherwise manually achieved for individual
+  component instances by sending changes via `LiveView.send_update/3` such an approach would not suffice in cases with
+  multiple component changes and more importantly, in the cases where there are elements required for insertion to or
+  deletion from the container list.
 
   The module relies on the reserved `:updated` key in the 'listilled' assigns, so developers need to make sure they
-  don't use it for other purposes when constructing assigns in the `c:PhoenixLiveViewExt.Listilled.construct_assigns/2`
-  callback implementation.
+  don't use it for other purposes when constructing assigns in their `c:PhoenixLiveViewExt.Listilled.construct_assigns/2`
+  callback implementations.
 
-  Below we use an example to explain step by step how to use Listiller. We do so on a non-trivial example of a table
+  Below we use an example to explain how to use Listiller step by step. We do so on a non-trivial example of a table
   (two dimensions of components - the row components and the cell components nested in them).
 
   ## TableView example
 
   Our example relies on the following assumptions:
   - `TableLive` module prepares the assigns for the corresponding `table_live.html.leex` template which lists all table
-    rows. The assigns include a temporarily assigned list of rows to append/update that is mapped to `:rows`. The list
-    contains only the rows that have actually changed and are to be treated as inserted, updated or deleted.
+    rows. The assigns include a temporarily assigned list of rows to append/update mapped to `:row_list_assigns`. The
+    list contains only the rows that have actually changed and are to be treated as inserted, updated or deleted.
   - `RowComponent` module prepares the assigns for the corresponding `row_component_t.html.leex` template which lists
     all cells in the row. Analogously, this involves the assigns used by the template itself as well as a temporarily
     assigned list of assigns for each cell that has actually changed (again, as in inserted, updated or deleted cells).
@@ -40,20 +40,21 @@ defmodule PhoenixLiveViewExt.Listiller do
     based on their diffing that the LiveView framework decides on what to send over the wire and what not.
   - Since an appended container in LiveView can only have its elements updated or appended, the final changes are
     performed in JS when we are sure that LiveView has applied all the necessary patches. The finalization involves
-    deletion of the marked-to-delete updated elements and sorting (repositioning) of the marked-to-sort appended
+    deletion of the marked-for-deletion updated elements and sorting (repositioning) of the marked-for-sorting appended
     elements.
 
-  The first step is to write our `TableLive` module. Note that the `:rows` assigns have to be constructed and assigned
-  every time the data may have possibly changed, involving all callbacks such as `handle_info` and `handle_event`.
+  The first step is to write our `TableLive` module. Note that the `:row_list_assigns` must be constructed and
+  assigned each time the data may have possibly changed, involving all callbacks such as `handle_info` and
+  `handle_event`.
 
       defmodule TableLive do
         use TableAppWeb, :live_view
-        alias PhoenixLiveViewExt.Listiller
+        alias PhoenixLiveViewExt.{ Listiller, Listilled}
 
         # If relying on handle_params to initialize then the initialization should be done there instead.
         @impl true
         def mount( _params, _session, socket) do
-          { :ok, initialize( socket), temporary_assigns: [ rows: []]}
+          { :ok, initialize( socket), temporary_assigns: [ row_list_assigns: []]}
         end
 
         defp initialize( socket) do
@@ -79,10 +80,9 @@ defmodule PhoenixLiveViewExt.Listiller do
         end
 
         defp construct_rows( socket, old_state) do
-          { rows, table_update} = Listiller.apply( RowComponent, old_state, socket.assigns)
-          socket
-          |> assign( :rows, rows)
-          |> assign( :table_update, table_update)
+          list_data = Listiller.apply( RowComponent, old_state, socket.assigns)
+
+          Listilled.Helpers.assign_list( socket, list_data)
         end
       end
 
@@ -93,20 +93,20 @@ defmodule PhoenixLiveViewExt.Listiller do
         import PhoenixLiveViewExt.Listilled.Helpers, only: [ phx_update: 1]
       end
 
-  .. and we make sure we properly render rows in our `table_live.html.leex`:
+  .. and we make sure we properly render the rows in our `table_live.html.leex`:
 
       ..
       <div id="table"
-        phx-update="<%= phx_update( @table_update) %>"
+        phx-update="<%= phx_update( @row_list_update) %>"
         phx-hook="TableHook">
-        <%= for row_assigns <- @rows do %>
+        <%= for row_assigns <- @row_list_assigns do %>
           <%= live_component @socket, RowComponent, row_assigns %>
         <% end %>
       </div>
       <div id="update-handler"
            class="hidden"
            data-print="<%= update_print( assigns) %>"
-           phx-hook="UpdateHook"></div>
+           phx-hook="UpdateHook"/>
       ..
 
   Note that we place the `update-handler` element _after_ the `table` element in our template. This is because we need
@@ -121,12 +121,12 @@ defmodule PhoenixLiveViewExt.Listiller do
   and is out of the scope of this library. The point is having the `UpdateHook`'s `updated` callback invoked once all
   elements in the lists have been patched by LiveView.
 
-  Now we provide the `RowComponent` module which, on one side is a `Listilled` `behaviour` implementation in charge of
-  rendering its templates and on the other, it provides the necessary behavior to support its nested `CellComponents`.
+  Now we provide the `RowComponent` module which on one side is a `Listilled` `behaviour` implementation in charge of
+  rendering its templates, and on the other it provides the necessary behavior to support its nested `CellComponents`.
 
       defmodule RowComponent do
         use Phoenix.LiveComponent
-        alias PhoenixLiveViewExt.Listiller
+        alias PhoenixLiveViewExt.{ Listiller, Listilled}
 
         ########################
         # Cell container related
@@ -134,25 +134,26 @@ defmodule PhoenixLiveViewExt.Listiller do
 
         @impl true
         def mount( socket) do
-          { :ok, socket, temporary_assigns: [ cells: nil]}
+          { :ok, socket, temporary_assigns: [ cell_list_assigns: nil]}
         end
 
         # Temporarily assigns the list of cell assigns for cell_components.
         @impl true
         def update( new_assigns, socket) do
-          { cells, row_update} = Listiller.apply( CellComponent, socket.assigns, new_assigns)
-          { :ok, socket
-                 |> assign( new_assigns)
-                 |> assign( :cells, cells)
-                 |> assign( :row_update, row_update)
-          }
+          list_data = Listiller.apply( CellComponent, socket.assigns, new_assigns)
+
+          socket =
+            socket
+            |> assign( new_assigns)
+            |> Listilled.Helpers.assign_list( list_data)
+
+          { :ok, socket}
         end
 
         ####################################
         # Listilled behaviour implementation
         #
 
-        alias PhoenixLiveViewExt.Listilled
         @behaviour Listilled
 
         # Prepares the row dom_id list and (optionally) supplements the provided state with any
@@ -169,7 +170,9 @@ defmodule PhoenixLiveViewExt.Listiller do
 
         # Returns dom_id as component_id
         @impl true
-        def component_id( dom_id), do: dom_id
+        def component_id( dom_id, _state) do
+          dom_id
+        end
 
         # Constructs new row assigns from the provided model state.
         @impl true
@@ -187,7 +190,7 @@ defmodule PhoenixLiveViewExt.Listiller do
         # Template helpers
         #
 
-        import PhoenixLiveViewExt.Listilled.Helpers
+        import Listilled.Helpers
         require PhoenixLiveViewExt.MultiRender
         @before_compile PhoenixLiveViewExt.MultiRender
 
@@ -205,17 +208,17 @@ defmodule PhoenixLiveViewExt.Listiller do
   Note that above we rely on the first available feature in the PhoenixLiveViewExt library v1.0.1 - the `MultiRender`
   `before_compile` macro that lets us define multiple (conditional) templates per LiveComponent.
 
-  Also, take note that our `RowComponent` module imports both `Distiller.Helper` functions because its component's
+  Also, take note that our `RowComponent` module imports both `Listiller.Helper` functions because its component's
   element requires post-append sorting while the template itself nests `CellComponents` thus requiring the `phx-update`
-  attribute set relative to whether to `replace` or or to `append` the cells that have changed.
+  attribute set relative to whether to `replace` or to `append` the cells that have changed.
 
   Below is the relevant part of our `row_component_t.html.leex` template:
 
       <div id="row-<%= @id %>"
            data-sort="<%= updated_sort( @updated) %>"
-           phx-update="<%= phx_update( @row_update) %>"
+           phx-update="<%= phx_update( @cell_list_update) %>"
            phx-hook="RowHook">
-        <%= for cell_assigns <- @cells do %>
+        <%= for cell_assigns <- @cell_list_assigns do %>
           <%= live_component( @socket, CellComponent, cell_assigns) %>
         <% end %>
       </div>
@@ -229,7 +232,7 @@ defmodule PhoenixLiveViewExt.Listiller do
 
         # Returns cell component id based on unique cell coordinates (a row dom_id and a cell key).
         @impl true
-        def component_id( { dom_id, key}) do
+        def component_id( { dom_id, key}, _state) do
           "#{ dom_id}-#{ key}"
         end
 
@@ -284,140 +287,126 @@ defmodule PhoenixLiveViewExt.Listiller do
          <!-- cell content -->
       </div>
 
-  Finally, we need some JS code to make everything work as intended.
-  The intention of the code below is _not_ to
-  provide a ready-to-use, copy-paste implementation (as neither is for any of the code in this documentation), but to
-  get a glimpse of the steps required for everything to function properly. However, the flow of invocation and the
-  sorting algo logic should remain the same.
+  Finally, there is a file with JS helper functions shipped with our library. To import it in your LiveView app,
+  simply add the following import before your JS code (app.js or the imported JS file therein):
 
-      let _sortElements = [];
+      import {
+        newListill,
+        prepForSorting,
+        completeListill,
+        applyCall,
+        initApplyCall,
+        deinitApplyCall
+      } from "../../deps/phoenix_live_view_ext/assets/js/listiller";
+
+  The sample code below relies on the imported JS helper functions.
 
       export const TableHook = {
         mounted: function() {
-          // adding listeners
-          this._afterModelUpdate = e => afterModelUpdate( this, e);
-          this.el.addEventListener( "afterModelUpdate", this._afterModelUpdate);
-        },
-        destroyed: function() {
-          // removing listeners
-          if( this._afterModelUpdate) {
-            this.el.removeEventListener( "afterModelUpdate", this._afterModelUpdate);
-          }
-        },
-        updated: function() {
-          this._isModelUpdated = true;
+          initHookState( this,
+            {
+              listill: newListill( 'sort', 'div[data-delete]')
+            });
+          initApplyCall( this);
         }
       };
 
       export const UpdateHook = {
-        updated: function() {
-          const event = new Event( 'afterModelUpdate');
-          document.getElementById( 'table').dispatchEvent( event);
-        }
-      };
-
-      function afterModelUpdate( me, e) {
-        const isModelUpdated = me._isModelUpdated;
-        me._isModelDisabled = false;
-        if( !isModelUpdated) return;
-
-        completeListill( me);
-      }
-
-      /* Completes the listill process by deleting all elements marked for deletion
-       * and sorting the inserted ones.
-       */
-      function completeListill( me) {
-        deleteElements( me);
-        sortElements();
-      }
-
-      /* Iterates over the FILO element array prepared for sorting (starting from
-       * the last pushed one) and repositions each of the elements found therein.
-       */
-      function sortElements() {
-        _sortElements.forEach( sort => {
-          const src = elementById( sort.src);
-          const dst = elementById( sort.dst);
-          if( src && dst) {
-            dst.parentNode.insertBefore( src, dst);
-          }
-        });
-        _sortElements = [];
-      }
-
-      /* Deletes all the elements tagged for deletion.
-       */
-      function deleteElements( src) {
-        src.el.querySelectorAll( 'div[data-delete]').forEach( el => {
-          el.dispatchEvent( new Event( 'killListeners'));
-          el.parentNode.removeChild( el);
-        })
-      }
-
-      /* Depending on row component/element implementation, the updated callback
-       * may or may not be needed in this Hook; same for CellHook
-       */
-      export const RowHook = {
         mounted: function() {
-          this._killListeners = () => {
-            if( this._killListeners) {
-              this.el.removeEventListener( "killListeners", this._killListeners);
-            }
-          };
-          this.el.addEventListener( "killListeners", this._killListeners);
-          prepForSorting( this.el, getRowSortId);
+          applyCallToTable( data => setTable( this, data));
+          completeTableListill( this);
         },
         updated: function() {
-          prepForSorting( this.el, getRowSortId);
+          completeTableListill( this);
         },
         destroyed: function() {
-          if( this._killListeners) this._killListeners();
+          deinitApplyCall( _table( this));
         }
       };
 
-      function getRowSortId( bareId) {
-        return elemId ? ( 'row-' + bareId) : null;
+      function completeTableListill( me) {
+        completeListill( _listill( _table( me)), _table( me).el);
       }
+
+
+      export const RowHook = {
+        mounted: function() {
+          applyCallToTable( data => setTable( this, data));
+          prepRowForSorting( this);
+        },
+        updated: function() {
+          prepRowForSorting( this);
+        }
+      };
+
+      function prepRowForSorting( me) {
+        return prepForSorting( _listill( _table( me)), me.el, getRowSortId);
+      }
+
+      function getRowSortId( bareId) {
+        return bareId ? ( 'row-' + bareId) : null;
+      }
+
 
       export const CellHook = {
         mounted: function() {
-          this._killListeners = () => {
-            // remove other CellHook listeners here
-            if( this._killListeners) {
-              this.el.removeEventListener( "killListeners", this._killListeners);
-            }
-          };
-          this.el.addEventListener( "killListeners", this._killListeners);
-          prepForSorting( this.el, getCellSortId);
+          applyCallToTable( data => setTable( this, data));
+          prepCellForSorting( this);
         },
-        destroyed: function() {
-          if( this._killListeners) this._killListeners();
+        updated: function() {
+          prepCellForSorting( this);
         }
       };
 
-      function getCellSortId( bareId) {
-        return 'cell-' + bareId;
+      function prepCellForSorting( me) {
+        return prepForSorting( _listill( _table( me)), me.el, getCellSortId);
       }
 
-  An important note: It is best advised that both sorting and deletion be done like this - in a bulk fashion all
-  at once after all the list elements have been rendered. The attempts to do it (e.g. deletion) on element by element
-  basis upon each element getting updated (i.e. interleaved with LiveView DOM patching) ends up in certain but
-  relatively difficult to reproduce bugs involving inconsistent DOM state.
+      function getCellSortId( bareId) {
+        return bareId ? ( 'cell-' + bareId) : null;
+      }
+
+
+      function initHookState( me, state) {
+        me._vars = state;
+      }
+
+      function _table( me) {
+        return me._tableObj;
+      }
+
+      function setTable( me, tableObj) {
+        me._tableOb = tableObj;
+      }
+
+      function _listill( me) {
+        return me._vars.listill;
+      }
   '''
   alias PhoenixLiveViewExt.Listilled
 
-  @typep assigns() :: Listilled.assigns()
-  @typep state() :: Listilled.state()
+
+  @type listilled() :: module()
+  @type list_update() :: :full | :partial
+  @type list_meta() :: %{
+                         listilled: listilled(),
+                         update: list_update(),
+                         version: Listilled.state_version()
+                       }
+  @type state() :: Listilled.state()
+
+
+  @typep script() :: [ { :eq | :ins | :del, list()}]
   @typep payload() :: %{
                         old_state: state(),
                         new_state: state(),
-                        inserted: MapSet.t()
+                        inserted: MapSet.t(),
+                        version: Listilled.state_version()
                       }
   @typep diff_id() :: any()  # must be unique relative to the diffs
-  @typep diff_insert() :: { :insert, { dst_id :: diff_id(), assigns()}}
-  @typep diff_delete() :: { :delete, assigns()}
-  @typep diff_update() :: { :update, assigns()}
+  @typep diff_insert() :: { :insert, { Listilled.sort_data() | nil, map()}}
+  @typep diff_delete() :: { :delete, map()}
+  @typep diff_update() :: { :update, map()}
   @typep diff() :: diff_insert() | diff_delete() | diff_update()
 
 
@@ -429,7 +418,13 @@ defmodule PhoenixLiveViewExt.Listiller do
   end
 
   @doc """
-  Distills the assigns for the components handled by the provided module that implements Listilled behaviour.
+  Distills a list of assigns for the provided component module implementing the Listilled behaviour.
+  Returns a tuple with the list of assigns and the list meta data. The returned tuple may be passed as such to the
+  Listilled.Helpers.assign_list/2 function to have the list assigns, the update type and the state version
+  assigned to the socket.
+
+  Raises ArgumentError if while diffing a `Listilled.component_id/2` function implementation returns a string value
+  containing the ':' character.
 
   Note:
   Due to the lack of access to the LiveView internally held diff state and the fact that we intentionally assign
@@ -439,14 +434,15 @@ defmodule PhoenixLiveViewExt.Listiller do
   (persistent) assigns to reduce memory load on the LiveView instance for it is expected that most if not all of the
   state supplied to the function is already kept stored with in the LiveView instance.
   """
-  @spec apply( module(), state(), state()) :: { [ assigns()], :full | :partial}
+  @spec apply( listilled(), state(), state()) :: { [ Listilled.assigns()], list_meta()}
   def apply( listilled, old_state, new_state) do
-    if !function_exported?( listilled, :state_changed?, 2) or
-         listilled.state_changed?( old_state, new_state)
-      do
+    old_version = Listilled.get_version( listilled, old_state)
+
+    if !function_exported?( listilled, :state_changed?, 2) or listilled.state_changed?( old_state, new_state) do
       { old_ids, old_state} = listilled.prepare_list( old_state)
       { new_ids, new_state} = listilled.prepare_list( new_state)
       list_update = old_ids == [] && :full || :partial
+      new_version = old_version + 1
 
       assign_list =
         listill(
@@ -455,22 +451,40 @@ defmodule PhoenixLiveViewExt.Listiller do
           %{
             old_state: old_state,
             new_state: new_state,
-            inserted: MapSet.new()
+            inserted: MapSet.new(),
+            version: new_version
           },
           []
         )
         |> then( fn { _, _, diffs} -> list_assigns( diffs, list_update) end)
 
-      { assign_list, list_update}
+      list_meta =
+        new_list_meta( listilled, list_update, assign_list != [] && new_version || old_version)
+
+      { assign_list, list_meta}
     else
-      { [], :partial}
+      list_meta = new_list_meta( listilled, :partial, old_version)
+
+      { [], list_meta}
     end
   end
+
+  # Instantiates a new list_meta map
+  @spec new_list_meta( listilled(), :full | :partial, non_neg_integer()) :: list_meta()
+  defp new_list_meta( listilled, list_update, version) do
+    %{
+      listilled: listilled,
+      update: list_update,
+      version: version
+    }
+  end
+
 
   # Recursively goes through a myers difference of row_id and by first collecting all diff_ids to be inserted
   # ensures those are not deleted as they are considered moved.
   # Relies on the gets_old_assigns and gets_new_assigns to fetch the assigns data and compare and them if different.
-  @spec listill( Keyword.t(), module(), payload(), [ diff()]) :: { diff_id(), MapSet.t(), [ diff()]}
+  # Raises if at an element insertion Listilled.component_id/2 returns a string value containing a ':' character.
+  @spec listill( script(), listilled(), payload(), [ diff()]) :: { diff_id(), MapSet.t(), [ diff()]}
   defp listill( [], _, args, diffs) do
     { nil, args.inserted, diffs}
   end
@@ -487,8 +501,7 @@ defmodule PhoenixLiveViewExt.Listiller do
         diffs -> [ { :update, new_assigns} | diffs]
       end
 
-    [ dst_id | _] = eq
-    { dst_id, inserted, diffs}
+    { List.first( eq), inserted, diffs}
   end
 
   defp listill( [ { :ins, ins} | rest], listilled, args, diffs) do
@@ -498,8 +511,15 @@ defmodule PhoenixLiveViewExt.Listiller do
     { dst_id, diffs} =
       for diff_id <- Enum.reverse( ins), reduce: { dst_id, diffs} do
         { dst_id, diffs} ->
+          sort_data =
+            if dst_id do
+              component_id = validate_id( listilled.component_id( dst_id, args.new_state))
+              { component_id, args.version}
+            end
+
           assigns = listilled.construct_assigns( args.new_state, diff_id)
-          diff = { :insert, { dst_id && listilled.component_id( dst_id, args.new_state), assigns}}
+          diff = { :insert, { sort_data, assigns}}
+
           { diff_id, [ diff | diffs]}
       end
 
@@ -521,16 +541,26 @@ defmodule PhoenixLiveViewExt.Listiller do
     { dst_id, inserted, diffs}
   end
 
+  @compile { :inline, validate_id: 1}
+  # Ensures the component id string does not contain a ':' character or raises ArgumentError if it does.
+  defp validate_id( id) when is_bitstring( id) do
+    unless id =~ ":" do
+      id
+    else
+      raise ArgumentError, "Component id (#{ inspect( id)}) contains a ':' character."
+    end
+  end
+
 
   # Returns the list of component assigns extracted from the supplied diffs.
-  @spec list_assigns( [ diff()], :full | :partial) :: [ assigns()]
+  @spec list_assigns( [ diff()], :full | :partial) :: [ Listilled.assigns()]
   defp list_assigns( diffs, list_update) do
     Enum.map( diffs, &diff_assigns( &1, list_update))
   end
 
   # Returns component assigns with sorting destination where :partial :insert
   # or :delete instruction where deleted.
-  @spec diff_assigns( diff(), :full | :partial) :: assigns()
+  @spec diff_assigns( diff(), :full | :partial) :: Listilled.assigns()
   defp diff_assigns( { :delete, assigns}, _) do
     Map.put( assigns, :updated, :delete)
   end
@@ -547,8 +577,8 @@ defmodule PhoenixLiveViewExt.Listiller do
     noop_assigns( assigns)
   end
 
-  defp diff_assigns( { :insert, { dst_id, assigns}}, :partial) do
-    Map.put( assigns, :updated, { :sort, dst_id})
+  defp diff_assigns( { :insert, { sort_data, assigns}}, :partial) do
+    Map.put( assigns, :updated, { :sort, sort_data})
   end
 
   defp noop_assigns( assigns) do